Sorting out a runable benchmark after merging the two branches
This commit is a combination of two branches (using decorated particles and benchmark), based on which we can do a benchmark for state-related cases so that it could be used elsewhere. It also enables two types of input (state or just positions Vector{SVector{3}(...)}). By doing so, I believe both the `benchmark` and `temp_benchmark_and_state` branches can be removed.